Provider of data instead of pagination

+++ src/main/java/org/krzywanski/table/table/DataProviderInterface.java
@@ -7,7 +7,7 @@ public interface DataProviderInterface<T> {
 
     List<T> getData(int limit, int offest);
 
-    long getSize();
+    int getSize();
 
-    long getLimit();
+    int getLimit();
 }
